Futher bug fixes

+++ backend/src/main/java/com/stasa/services/MemberService.java
@@ -17,7 +17,7 @@ public class MemberService {
         return memberRepo.findAll();
     }
 
-    public List<Member> getByUserId(int userId) {
+    public List<Member> getByUserId(long userId) {
         return memberRepo.getByUserId(userId);
     }
